fix: solhint fix, explicit types

+++ contracts/src/v0.8/ccip/capability/CCIPConfigArraysValidation.sol
@@ -17,8 +17,8 @@ library CCIPConfigArraysValidation {
         _checkSortedAndDuplicates(b);
 
         // Check if 'a' is a subset of 'b'
-        uint i = 0; // Pointer for 'a'
-        uint j = 0; // Pointer for 'b'
+        uint256 i = 0; // Pointer for 'a'
+        uint256 j = 0; // Pointer for 'b'
 
         while (i < a.length && j < b.length) {
             if (a[i] > b[j]) {
@@ -40,7 +40,7 @@ library CCIPConfigArraysValidation {
 
     // Helper function to check if array is sorted and has no duplicates
     function _checkSortedAndDuplicates(bytes32[] memory array) private pure {
-        for (uint i = 1; i < array.length; ++i) {
+        for (uint256 i = 1; i < array.length; ++i) {
             if (array[i] < array[i - 1]) {
                 revert NotSorted(array);
             }
